requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
astroid==2.14.2
attrs==23.2.0
beautifulsoup4==4.11.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comments==1.0.0
deepdiff==6.3.1
dill==0.3.6
Django==5.0.6
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-modal-forms==2.2.0
django-bootstrap-v5==1.0.11
django-comments-xtd==2.10.0
django-contrib-comments==2.2.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-file-resubmit==0.5.2
django-htmx==1.14.0
django-mathfilters==1.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.15.1
docutils==0.21.2
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.5
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
ordered-set==4.1.0
pandas==2.0.1
Pillow==9.4.0
platformdirs==3.0.0
pydantic==1.10.5
PyJWT==2.8.0
pylint==2.16.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2023.3
reportlab==3.6.13
requests==2.31.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomlkit==0.11.6
twilio==9.0.5
typing_extensions==4.5.0
tzdata==2022.7
urllib3==2.0.4
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.1.0
yarl==1.9.4
